Remove optional dependencies
We do not foresee this Helm chart being used with consul or zookeeper.
Also, we do not want to strictly rely on etcd, as we can always run with
native Kubernetes support.

Therefore, remove the requirements of these and the optional
configuration, in order to have a more focussed set of templates.

Also, WAL-e is not our preferred backup solution, so removing those
configuration items as well.