Enable pgBackRest backups to be made
The pgBackRest backups can be created by triggering a call to a tiny
wrapper around pgBackRest.
To be able to create backups pgBackRest requires some credentials to be
available. This first attempt will use Kubernetes secrets that contain
AWS key ID and an AWS Key Secret to store the backups.

To ensure both scenario's work (with and without backup) some evaluation
of the PostgreSQL parameter archive_command is needed: Without it being
set properly the backup will fail - which in the worst case means we
don't have a backup, but even in less that worse case scenario's this
will cause the WAL disk to fill up.

If we therefore find that the backup is not enabled, we explicitly set
this command to an always succeeding value. Another option is to disable
`archive_mode`, however, that requires a PostgreSQL restart, whereas
changing `archive_command` only requires a reload. The latter therefore
has less impact if someone decides to switch to using a backup later on.

We currently don't schedule any backups yet, that should also be part of
this Helm chart. The most likely used Docker image will likely trigger
an initial backup after initdb, but we cannot rely on that to have sane
backups.

Part of this commit moved the literaly included SPILO_CONFIGURATION into
its own high level `patroni` key. This ensures we can use the
configuration in our templates, to replace, override or remove if
necessary.