Add CronJobs to trigger backups
For every configured job, iff backup is enabled, we schedule a
kubernetes CronJob.

I chose to use a list of Jobs to allow full expressiveness for anyone
using this Helm chart. With this approach we should be able to support
pretty much every backup schedule someone may come up with.