Pass on metadata labels to Jobs and Pods
Adding the metadata labels to not just the CronJobs, but also the Jobs
it creates and subsequently the Pods that are created to fullfil the job
allows us to easily identify all the objects belonging to a release.

As an example, to list all the jobs that are part of a release, after
this commit we can do the following:

    kubectl get job -l release=my-release

whereas previously we were condemned to using grep or other filtering to
find the jobs we're interested in.