Allow most PostgreSQL Docker images to be used
This Helm chart should be able to use almost any Docker Image that has
roughly the same design as the postgres Docker image from Docker hub.
To ensure they all work however, we need to pass on some environment
parameters explicitly, like PGDATA.

The first use case that is fulfilled by this is the ability to create a
Docker image from the github.com/timescale/timescaledb using the
currently available script, and use that created Docker image for your
deployment.