Separate secrets for Access node and Data nodes
A data node does not need to connect to an Access Node. If the
credentials are shared however, the Data Node has access to the Access
Node.
By splitting the secrets we should ensure that even a superuser on a
Data Node cannot connect to an Access Node as it does not know the
password for the Access Node.