Express Access/Data Nodes individually
In order for us to add High Availability and some future improvements we
are going to be deploying a cluster as a set of StatefulSets, 1 each for
each Node in the topology.

As we still want to expose configuration items to the deployer and we do
not want operators to have to repeat themselves, we introduce the
concept of NodeTemplates, which can be referenced in the topology key.

This should aid in focussing on the high-level architecture when
specifying the topology, but not taking away the option to fully
override each and every single value.

We use anchors and aliases[1], togheter with the << operator[2], which
means none of this has anything to do with Helm, but this is purely a
YAML way of repeating certain values.

https://yaml.org/spec/1.2/spec.html
https://yaml.org/type/merge.html