Fully configure Patroni from kubernetes
The SPILO_CONFIGURATION environment variable allowed Spilo[1] or our
own Docker image[2] to call the configure_spilo.py script.
This python script sets up Patroni and does it correctly, however the downside
is that it is quite opaque at what is happening when running this in a
Kubernetes context.
Secondly, it makes it harder to allow Operators - whether human or
automated - to express explicit Patroni configuration. The previous
implementation caused the configuration to be touched and transformed in
many ways.

Part of the work done was to remove all spilo-isms, e.g. the label
spilo-role is no longer used in favour of role.

With this commit in, we can
- fully control Patroni (and therefore PostgreSQL parameters)
- fully control pg_hba
- inject SSL certificates via the values
- override clusterName

This also means we do not rely on an entrypoint anymore for Patroni,
however we need some more work for pgBackRest before we are free of any
entrypoints.

[1]: https://github.com/zalando/spilo
[2]: https://github.com/timescale/timescaledb-docker-ha