Configure pgBackRest fully from k8s objects
Instead of using a container runtime script that translates some
environment variables into a configuration file, we now create the
configuration file as a Secret (as it contains the passwords).

This removes the need to synchronize which container is doing what at
startup time (which container should write the configuration?), as the
configuration is now mounted in both the timescaledb as well as the
pgBackRest container.

To further decouple the Helm chart from the Docker image used, we also
no longer rely on the existence of a specific entrypoint in the Docker
image, we write our own bootstrap script.

We are still relying on the backup wrapper (written in Python) to exist
in the Docker image we use.