Bugfix: Tell the LoadBalancer which port to follow
The Service and Endpoint port names need to match, see:
https://patroni.readthedocs.io/en/latest/SETTINGS.html#kubernetes

Part of this is the renaming of clusterName; the name does not need to
be globally unique and therefore adding the namespace to the name is not
needed.

In this way we get some shorter, cleaner names for the objects
generated.