Use Namespace as a subfolder for s3 backups
Followup on 97faa7036981981f0414bcdc4e3d8b7640a2fc5c, where clusterName
was shortened not to include the namespace.

This is fine, as everything inside Kubernetes is namespaced and
therefore no clashes should exist, but for the s3 bucket containing the
backups this may cause name collisions.

This change ensures that different clusters in different namespaces will
end up in different folders in s3,