Allow WAL to be stored on its own Volume
To quote the PostgreSQL manual[1]:

   It is advantageous if the log is located on a different disk from the
   main database files. This can be achieved by moving the pg_wal
   directory to another location (while the server is shut down,
   of course) and creating a symbolic link from the original location in
   the main data directory to the new location.

With this commit running WAL on a different volume from the data volume
is trivial. No huge benefits will be available straight away, but with
this commit we can start exploring different storage classes for the
different volumes, e.g. experiment with local disks for WAL and EBS
Volumes for PGDATA.

The configuration of Patroni is the same with or without a separate
volume:
If no WAL volume is created, the WAL files are stored on the
same Volume as the data, and a symlink is created to that outside
directory. However, when a WAL Volume is specified, the symlink points
to a directory outside the Data Volume.
This allows us to migrate from a deployment without a separate WAL
volume to one with a separate WAL Volume: Rescheduling the pod should
suffice.

Switching the mountPaths from /home/postgres to /var/lib/postgresql
brings us back in line with Debian and the official Docker image of
PostgreSQL[2], the only reason we used /home/postgres is the heritage of
these Helm charts and their Spilo dependency.

We do not use these mountpoints directly, as that is deemed unwise,
again we can quote the PostgreSQL manual[3]:

   Many installations create their database clusters on file systems
   (volumes) other than the machine's “root” volume. If you choose to do
   this, it is not advisable to try to use the secondary volume's
   topmost directory (mount point) as the data directory. Best practice
   is to create a directory within the mount-point directory that is
   owned by the PostgreSQL user, and then create the data directory
   within that. This avoids permissions problems, particularly for
   operations such as pg_upgrade, and it also ensures clean failures if
   the secondary volume is taken offline.

The tools that work with the WAL directory directly (initdb,
pg_basebackup, pgBackRest) are also reconfigured to reflect this change.

A minor concern is that Patroni may not be fully capable of working with
a separate WAL directory, when it creates a replica and fails, it
removes PGDATA, but would leave the WAL directory intact. We will need
to fix this (probably upstream to Patroni) if this becomes a recurring
problem.

[1]: https://www.postgresql.org/docs/current/wal-internals.html
[2]: https://github.com/docker-library/postgres/
[3]: https://www.postgresql.org/docs/current/creating-cluster.html#CREATING-CLUSTER-MOUNT-POINTS