Bugfix: Always set wal directory, regardles of Volumes
Commit 188848d4111b77f98c7b3fa3878191d634981d06 introduced the
possibility to create separate WAL Volumes. With that change, the WAL
directory will now always be outside the PGDATA directory, even if the
Volume is the same.

We should therefore always set the wal_directory.

This came to light with error messages when running
`install [...] ""` where the wal_directory should have been inserted.

Backpatch to 0.2