Associate IAM role with a Service Account (AWS)
Currently, the pgBackRest backup requires secrets to do its backups to
S3. These secrets are stored as k8s secrets, however they are specified
as plain text in the values.yaml file.
By associating an IAM role with a Service Account we remove the need for
specifying these secrets in the values altogether.

https://docs.aws.amazon.com/eks/latest/userguide/iam-roles-for-service-accounts-technical-overview.html