Drop Timescale extension from the Data Nodes
During the process of bootstrapping Data Nodes from Access Nodes, there
are 2 options (currently):

bootstrap => false will create the database *and* install the extension
bootstrap => true  will create neither the database nor the extension

The problem with having TimescaleDB in the template1 database is that
after  CREATE DATABASE example is executed, it already contains a
TimescaleDB extension. This means bootstrap => true will fail, as it
cannot install the extension.

This is probably to be fixed in another way at some point, however for
now the safe option is to drop the extension from the template1
database.