Allow Load Balancer for the replica's
By always creating a service for both the primary and replica's, and
optionally attaching a Load Balancer to the service we achieve that
there is always a service to connect to *from inside kubernetes*.
One service for the primary and one service for the replicas.

If the [replica]LoadBalancer is also enabled, than we will also expose a
LoadBalancer, which in an AWS scenario would become an ELB/NLB, which
would allow connections from outside the kubernetes cluster to the
database.

In passing, it seems we never use the `timescaledb` ep, better to remove
it then.