Include readinessProbe for PostgreSQL by default
According to the Kubernetes Documentation[1], this is how pods are
rescheduled currently:

    ... updating each Pod one at a time. It will wait until an updated
    Pod is Running and Ready prior to updating its predecessor...

Without the readinessProbe the container running PostgreSQL was
considered Ready very quickly after startup. However, we can express
readiness much better by saying that it needs to *at least* be able to
connect to the PostgreSQL instance in the container before the container
can be marked ready.

This fixes a problem like the following scenario:

- An update to the Pod Template, requiring a rescheduling of all the
  pods
- A somewhat large - or active - database that requires significant time
  to be restored

What would happen is that a RollingUpdate strategy (which is the default
but with this commit is also made explicit), the first pod to be
restarted would be marked Ready way before it could even accept
PostgreSQL connections, let alone be a candidate for a master failover.

This readinessProbe is not perfect, as it only verifies if connections
are allowed. Not whether the StatefulSet as a whole can currently cope
with a failover.

That is not easily solved with only ReadinessProbes, as it requires a
distributed decision on whether or not this pod is healthy enough. That
is where a Helm Chart / StatefulSet / updateStrategy start to show its
limitations and where an operator can make the correct decisions.

[1]: https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#rolling-updates