Explicitly set PostgreSQL Environment Variables
Although PGDATA and PGHOST are not directly used by Patroni and
therefore by PostgreSQL, it is very convenient to have them defined
correctly.

In the case of PGDATA, chances are high that the value was actually
exposed incorrect previously, as many PostgreSQL Docker Images set the
PGDATA environment variable explicitly. This value is very likely to be
different from the Data directory that is used in this Helm Chart - even
more so as the Data directory can be configured directly by the user of
this chart.

By overriding PGDATA explicitly with the Data Directory we also pass on
to Patroni we ensure PGDATA is correct. Some immediate advantages:

- Executing a command that may rely on PGDATA just works (think:
  pg_controldata)
- Callbacks called by Patroni have their environment setup correctly