Configure socket permissions in PostgreSQL
Previously, we attempted to restrict permissions of the socket by
setting `defaultMode` on the EmptyDir that serves the PostgreSQL socket.

This however is not a valid configuration option for EmptyDir[1], and
caused some issues when the Chart was deployed using for example Helm
2.14.0:

        Error: validation failed: error validating "":
        error validating data: ValidationError(StatefulSet.spec.template.spec.volumes[0].emptyDir):
        unknown field "defaultMode" in io.k8s.api.core.v1.EmptyDirVolumeSource

By configuring the permissions using PostgreSQL parameters we achieve
what we set out to do: protect the socket for every user, and now we
actually do it correctly.

Addresses issue #57.

1: https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.16/#emptydirvolumesource-v1-core