Initialize replica using the pgBackRest backup
Previously, the only method available to a replica was to initialize
using pg_basebackup. By adding the method to initialize using pgBackRest
the replica does not put any extra load on the primary node, as it
fetches everything it needs from s3.

The downside is also that it will fetch everything from s3; if the
backup is 4 days old, the replica will fetch 4 days worth of WAL to
catch up with the primary.

By also allowing this process to be done without a master being present
allows reuse of a previous deployment name - it will be restored from
the backup in that case.

If passwords differ between the previous deployment and this deployment,
there may still be some issues with connecting a standby to a master.