Change Wal Volume Size from 1G to 1Gi
The trigger for this update was a Digital Ocean deployment, which
requires volumes of at least 1Gi.

Update the max_wal_size/min_wal_size default parameters as well, to
ensure the wal volume can actually be used. The choice of 750MB is to
ensure the disk usage (by default) will remain below 80%, which is a
threshold that is very commonly configured in monitoring for alerts