Switch services to ClusterIP if LB is disabled
Previously we used a headless service if the Load Balancer (LB) was
disabled for a given service. This however runs into an issue[1] which
cannot be easily solved without having to do manually
patch/remove/update some kubernetes objects.

To allow an easy path for upgrading a non-LB service to a LB service, we
change the service from being headless to being a ClusterIP service.
This allows users of the charts to upgrade by flipping `enabled=False`
to `enabled=True` for the service they want.

Changing the service from a LoadBalancer to a ClusterIP however does not
work currently, and requires some manual steps. As we cannot currently
solve this easily in the Helm chart we document the steps that allow
this upgrade to take place.

1: https://github.com/kubernetes/kubernetes/issues/11237