Make Prometheus scrapeable and add more config
Previously, even if you enabled Prometheus support, it was not
scrapeable because there was no service with the appropriate
annotations to tell Prometheus how to find it. To fix this, we now
deploy a service if Prometheus is enabled.

Further, to support more configuration of the Prometheus Postgres
exporter, we add the ability to specify env vars for specifically
the exporter container as well as volumes. This allows one to do
things like add custom queries to have their results exported with
other Prometheus metrics.