Create directories before restore if missing
In some scenario's it can be that the PGDATA directory is missing,
Patroni will remove the PGDATA directory for example if it is unable to
bootstrap a replica correctly.

When this situation occurs, pgBackRest will refuse to commence the
restore with the following error message:

        ERROR: [073]: $PGDATA directory '/var/lib/postgresql/data' does not exist

Even though this can happen, it is not the normal sequence of events:
the PGDATA directory is explicitly created on container start.

If PGDATA does not exist it has to have been either Patroni or a (human)
operator that has removed this directory.

If Patroni removes PGDATA, it will also remove the WALDIR[1], and therefore
we will also have to recreate that sometimes.

1: https://github.com/zalando/patroni/blob/v1.6.3/patroni/postgresql/__init__.py#L782-L789