Add support for NetworkPolicies
For Kubernetes clusters that make use of NetworkPolicy and CNI, we
should provide a basic policy that they can extend with additional
rules.

To that end, this PR adds support for such a policy that makes sure
that all pods within the app can talk to one another (i.e. for
replication), can talk on the Patroni multicast channel for leader
races, and can communicate with backup scripts. Additionally, we
also have an ingress rule to allow Prometheus scrapers to connect
to the Prometheus endpoints. These ingress rules can be extended
by setting new rules under the 'networkPolicy.ingress' variable.