Exclude png files from the Helm Release
Helm stores items from the Chart inside Kubernetes objects.
Helm 3 for example stores them as secrets.

ConfigMap and Secret objects have a real-life limit of 1MB in most
kubernetes clusters, because of a limit of etcd[1].

We're getting close to that limit, which is mostly because we add png
images to the releases. This commit ensures we do not do this anymore.

Before:
$ k get [...].v1 -o json | wc
     23      41  713256
After:
$ k get [...].v2 -o json | wc
     23      41   73678

1: https://stackoverflow.com/a/53015758/2459656