Fail fast on datadir or waldir issues
The install command should not fail if permissions are correct or if the
directories already exist. As a failure on the data directory or the wal
directory will also cause Patroni and PostgreSQL to fail, it seems
better to fail fast with the error message rather than to continue,
which will clutter the logs with more error messages.

For example in issue #72, the output would have been very clear that it
is a permission problem.