Expose Patroni through the -config Service
Previously this Service would have been created by Patroni and would be
created without a Selector. After this commit, this (headless) service allows
connections to the Patroni port.

As a side effect, this will also remove the -config Service and
Endpoints if a `helm delete` or `helm uninstall` is issued, due to the
Helm annotations. This means less manual work when purging a deployment
from Kubernetes.