Update Patroni DCS Configuration on Helm Update
Previously, changes in the `patroni.bootstrap.dcs` section of the
values.yaml would not propagate anywhere, as this section is only read
once (during initdb time), and then persisted in the DCS.

The documented way to update the dynamic configuration[1] is to:

1. update the DCS directly (kubectl edit ep/RELEASE-config)
2. or to send a PATCH to the /config endpoint[2] of Patroni

This commit implements option 2; it will connect to any running Patroni
to PATCH the config endpoint, for every update of the Helm Deployment.

The downside of option 1 is that it tries us to the DCS directly, whereas
Patroni can also use a separate etcd, consul etc, instead of relying on
Kubernetes endpoints, so it would turn a preferred option (use
endpoints) into the only option available in this Helm chart.

This addresses Issue #74

1: https://patroni.readthedocs.io/en/latest/dynamic_configuration.html
2: https://patroni.readthedocs.io/en/latest/rest_api.html#config-endpoint