Tune pod using timescaledb-tune
To ensure timescaledb-tune works well with a Patroni managed PostgreSQL
instance, we create a separate configuration file which is fully
generated by timescaledb.
We will then use the include directive of PostgreSQL configuration files
to load this file.

This means the order of evaluation of settings becomes as follows:

1. PostgreSQL compiled defaults
2. postgresql.example.conf from the package
3. timescaledb-tune generated file
4. Patroni: bootstrap.dcs.postgresql.parameters
5. Patroni: postgresql.parameters
6: Patroni: overrides (cluster-name for example)

This means that if `shared_buffers`, is not specified in the Patroni
configuration it will use the value of timescaledb-tune.

Timescaledb-tune is currently not very well suited to set min/max wal
size, so in order to tune those we do some shell arithmetic.

Disabled this feature by default, to allow some evaluation to take
place.