Do not label Job Pods with cluster-name
The cluster-name label is used by Patroni to identify which Pods are
part of the replication topology.

Up to this commit, we labelled every component with the `cluster-name`
label for ease of selection, however, this has a major downside in that
when a Pod is not running Patroni, but it does have the same labels as
set in the PATRONI_KUBERNETES_LABELS environment variables, it is
considered to be part of the replication topology by Patroni.

Patroni *will* therefore create a replication slot for that pod
- which will never be removed as it is not a member of the topology.

Stale replication slots cause WAL pile up and at some point will cause
the WAL volume to fill up, which may cause the PostgreSQL instance to
become unavailable.

As this issue is only happening for pods, I decided to only remove the
labels for the Job Pods, not for the CronJobs or Jobs themselves.

This addresses issue #106

1: https://github.com/zalando/patroni/blob/dc1966e3bcba9fde7f6113e72fdd97395aab6e32/patroni/dcs/kubernetes.py#L201