Include example for a High Throughput deployment
The example includes asynchronous archiving, synchronous commit and
large volume sizes. Additionally, as it wasn't that well documented yet,
it gives an example on how to reference an existing secret for the s3
credentials.

I'm not really fond of using ':' in yaml keys, but in this case it seems
the easiest way to map the sections of the pgBackRest configuration file
to yaml keys, without breaking previously built `values.yaml` files.

 pgBackRest              --> global
 pgBackRest:archive-push --> global:archive-push
 pgBackRest:archive-get  --> global:archive-get

These settings were tested on a cluster (3 pods on a dedicated 3 node
m5.xlarge ec2 nodepool) doing > 100K row inserts per second.

To make the example work correctly, some refactoring (backwards
compatible) of `sec-pgbackrest.yaml` was required, which also addresses
issue #109, which was raised to decouple the clusterName from the s3
backup repository path.