Disable replication slots (high throughput example)
Using replication slots in high throughput situations has a significant
danger: The master may need to retain more WAL than the size of the WAL
Volume allows, if a replica is catching up.

With backup enabled: while the replica is catching up it will be
fetching archives from s3, until s3 is exhausted, after which the
replica will switch to streaming from the master. While the replica is
not yet streaming (this may take hours), the replication slot reserved
for it is stale and will cause WAL to be retained by the master.  If we
have a backup, we don't need replication slots, as s3 will have the WAL,
therefore, if the backup is enabled, we can safely disable the use of
replication slots.