Improve speed of failover and recovery
By issuing a switchover *before* the pod is terminated, we avoid any
waits that may occur when the Pod is terminated while being a primary.
For relatively idle clusters, this brings the downtime of a primary pod
reschedule to something like 5 - 10 seconds.

For more active clusters, most of the time during shutdown is spent in
the final CHECKPOINT. However, the final CHECKPOINT is issued *after*
clients have already been disconnected, and no connections are allowed
anymore.
Very active clusters may take minutes to complete the CHECKPOINT, which
reduces the availability of the databases. By issuing a CHECKPOINT
*before* the actual shutdown commences, we may still take a significant
amount of time, but we establish 2 things:

1. Clients can still do their work, connect, read/write
2. The CHECKPOINT issued by the shutdown will be faster (if not very
   fast) as the previous CHECKPOINT was only moments ago.

As a restartpoint (CHECKPOINT for replica's) can only be written on the
CHECKPOINT record from the primary, we also issue a CHECKPOINT on the
primary if one of the replica's is terminating.
Recovery of the replica starts at the previous CHECKPOINT record, which
if we do not issue a CHECKPOINT, may be many minutes (say 10) in the
past.

For very active clusters (say 80% CPU usage for the recovery process),
this may mean that the replica will only catch up 1.25 minutes every
minute, or in other words, it requires 40 minutes to recover the 10
minutes from the previous CHECKPOINT.

If the CHECKPOINT issued here however is recent (say the pod restarts in
1 minute) we now only need 4 minutes to catch up with the primary and
resume the duty of a replica.

https://www.postgresql.org/docs/current/wal-configuration.html