Scrub the Patroni/PostgreSQL environment
Environment variables that are present at postmaster start may leak to
regular user sessions.
We therefore want to only pass along those environment variables that
are necessary for Patroni/PostgreSQL to function.

Patroni itself is aware of this, and therefore will already scrub all
the environment variables starting with PATRONI_ from the environment
before starting PostgreSQL, however, other environment variables may be
set that carry secrets, as that is quite a common setup in a Kubernetes
deployment.

To ensure the environment *is* available to callbacks, archive scripts,
backup scripts etc we store the environment before invoking Patroni and
we source the environment if any of these scripts are executed.

The immediate use case that is covered here is the exposing of S3 access
keys/secrets that are injected using `envFrom` or `env` settings in the
Values files, which in turn are referencing Kubernetes secrets.