Generate deployment specific secrets without Helm
We used to specify secrets using a credentials section in `values.yaml`,
however this has a few major drawbacks:

1. The literal passwords/certificates were part of the `values.yaml`, and
   available in a ConfigMap (Helm 2) or Secret (Helm 3) to Helm.
2. Default credentials are used, as it is not required to change these
   values before invoking Helm.
3. Secrets would be removed once Helm removed a deployment. (This can
   also be a good thing though). Recreating a deployment would therefore
   create a new set of secrets, which would not match those in case of a
   restore from backup.

This commit does not yet document this process yet, as the current way
of working is not removed yet, this is just one of the prerequisites to
start a new deployment.

Future commits will also remove the previous credentials support.

This addresses the following issues:
- #70 (partly)
- #112