Make patching of Patroni optional
The previous definition of the Job would be required to succeed before
Helm would be satisfied.
The downside of this however is that if no Patroni is running no Helm
upgrades can be done.

If no Patroni's are running, patching the configuration is also not
important, as there are no running instances that require an immediate
reload.