Initiate backup from the backup sidecare
Previously, the backup was only triggered after a database
initialization. As this is a one-time event, if the backup was
(temporarily) broken, or the deployment started out without a backup, no
initial backup would be created.
The first backup of that cluster would then be the first one triggered
by the CronJobs.

Every (restarted) pod that start PostgreSQL will have PostgreSQL running
in recovery. This is by design, and Patroni will promote the instance if
necessary.
This however caused quite a few issues where the backup sidecar did not
initiate a stanza-create as the database was still in this initial in
recovery phase.

A classic race, which is classically solved by sticking a sleep 10 in
there.