Use Docker Image PostgreSQL version by default
In the previous tag, the default was set to 11, however this was not
propagated to either the PATH environment variable or the
patroni.postgresql.bin_dir, causing the behaviour to be broken.

Testing shows that the recovery options that were part of the pgBackRest
configuration are not required even for PostgreSQL 11, so we can remove
those lines altogether.

We still do need the option for someone to specify the version
explicitly, this allows someone to switch to a PostgreSQL 12 Docker
Image, while still running PostgreSQL 11, this allows you to decouple
the upgrading of Docker Images from running the pg_upgrade.