Create random credentials once if unsafe == true
During development and quick evaluation of the Helm Charts it is useful
to not worry about credentials too much.

Currently, some pre-installation steps are required to install this
Chart successfully. The value of those steps is that it ensures the
Secrets that are used are outside the control (and values) of Helm.
However, this is not as easy as generating some random credentials.

Therefore, introduce the unsafe boolean, it will create some random
credentials and a random certificate before installation.