Include example kustomization in Helm package
The goal has been to never accidentally include non-example
kustomizations, yet not to forget the example kustomization.

This however does not work the way we want it as expressed in
issue #157.

The problem is that .helmignore files are broken when it comes to
negation[1], which we tried to address in commit 25ee3b95, but that
approach also seems to fail.

Instead of relying on .helmignore, we'll now rely on the Makefile to
verify that no non-example kustomizations are in the directory. It's not
pretty, but it works and includes the example sections.

1: https://github.com/helm/helm/issues/3622