Include option to run pgBouncer as a sidecar
pgBouncer[1] can be very useful in certain environments to limit the
number of connections established, or to limit the number of active
connections to the database.

There are quite a few deployment models for pgBouncer, and it boils down
to the following:

> Q: Should PgBouncer be installed on webserver or database server?
> A: It depends
https://www.pgbouncer.org/faq.html#should-pgbouncer-be-installed-on-webserver-or-database-server

This implementation is chosen, as other implementations can be done
outside of this Helm Chart. This implementation (a sidecar) however
requires to be integrated with the Helm Chart in order to work.

Addresses github issue #183

1: https://www.pgbouncer.org