Prevent WAL disk from filling up
One of the toughest situations to resolve is the WAL Disk running out of
space. To recover from this, you either have to reclaim space, or extend
the volume size and the filesystem.

Extending the volume size is non-trivial in a Kubernetes context, and
reclaiming space is very hard, as WAL segments should never be (re)moved
manually, as they are essential to crash recovery of the PostgreSQL
instance.

(Having the Data Volumes run out of disk space is not that big of an
issue, as PostgreSQL can keep running).

This script and feature adds the possibility to verify that we're not
running out of diskspace for every time the archiver is invoked. By
doing it at that point in time, we can ensure the check is not executed
needlessly frequently on idle clusters, but it does get invoked often
for very write-heavy clusters.

This means that this script - once it has seen some good testing -
should prevent most occurrences of running out of WAL space. Instead of
the PostgreSQL instance not being available at all anymore, it will
still be available for read-only queries, until the time that enough
disk space has been reclaimed. Once the PostgreSQL instance is running,
we can reclaim diskspace in a safe way, for example:
- drop stale replication slot
- fix archiver process
- issue checkpoint

With regards to the choice for bash here: a psql script with variables
would also work, but it will be harder to maintain, as the psql
conditional blocks (\if \endif) are hard to read and not seen that often
in the wild.
I also tried to use a Python script here, as that makes for easier
reading, however the major downside there was that starting the Python
interpreter is an order of magnitude slower than just using bash.