Run tests for all pushes
Introducing these tests gives us reasonable confidence that we're not
introducing a major error with the primary purpose of these Helm Charts:
to deploy a working set of Patroni managed PostgreSQL members.

The linter runs on every push, whereas the more extensive test, which
spins up all known values files, is only triggered if the Chart.yaml has
been changed in preparation of a release.