Refactors pgbouncer configmap
This is a minor refactor, putting the configmap values from the values.yaml at the END of the block instead of the beginning.
This allows for overriding of all options, specifically of importance is auth_query.

The reason for this is pgbouncer uses the last instance of a given configuration key, so in the orignial structure, adding an auth_query to the values.yaml resulted in that setting getting ignored, using the default from the template.
The same was true for all template-provided defaults.

It also adds override instructions and a warning to values.yaml