Refactor some pgBouncer features
Made the pgBouncer pg_hba into a list, as that is also how the Patroni
pg_hba is configured. It would be confusing if in the same `values.yaml`
there would be 2 separate ways of configuring the same category of
thing.

Added schema and documentation for these new pgBouncer configuration
items.