Use label selector to find pod with role=master
When run under Kubernetes, Patroni normally updates the Endpoints of a Service so the current-master node is discoverable by connecting to that Service.

When patroni.kubernetes.use_endpoints=false, this functionality is disabled, and nothing creates those Endpoints. This leaves the current-master difficult to discover.

However, both role=master and role=replica labels are updated on the pods. The role=replica label selector is already used in the `RELEASENAME-replica` Service. Use the role=master label selector in the `RELEASENAME` Service too, when patroni.kubernetes.use_endpoints=false.

Fixes issue #278