Remove failing workflow
As we can see in the repo, this workflow is being triggered without
conditions being 100% satisfied, so removing the workflow to avoid
having wrong executions on merges to master.

Also added a few notes of deprecation in multi-node chart files.
an remove project management workflow.
Tweaks chart-lint config to skip multinode and promscale.