Opt to enable linkerd proxy shutdown from lifecycle preStop
Added a new `linkerd` object to values & schema. If `enabled=true` &
`shutdownAfterCheckpoint=true`, then the STS pod's `lifecycle.preStop`
hook will be configured to issue a shutdown call to the local
linkerd-proxy.

This is entirely optional. If a deployment is not using linkerd, then
this option should be left in the default disabled state.