Use sidecar to monitor events on timescaledb container, then cleanup (hopefully last pattern to be tried)
Given a few peculiarities with how liveness probes work, and the fact
that the underlying timescaledb container could end up being removed by
the underlying kubelet before the liveness probe has a change to perform
its cleanup logic, it appears as though this pattern provides the most
robust approach.

The fact that we are monitoring the K8s API for events related to the
specific target container, and that we are using timestamps of the
events to maintain event stream offsets, should ensure that this pattern
is the most robust pattern for guaranteeing that a specific cleanup
routine is performed when the timescaledb container exits, all without
touching the timescaledb container itself.